Feature Delete: Purge.
Sphinx runs the purge event when a document changes. So there's no easy
way to determine if the function is called when a document is just
removed or merely changed. It was clearing cached data on still-existing
documents.

Removing this feature for now. If some Sphinx-enabled project uses tons
of Imgur references then I guess it might be a problem one day if they
keep the pickled file for long. For now I'm removing this feature.
Cached entries will be stored forever.